improve glslc invocation compatability with macOS
It's annoying to install `glslc` on macOS, but what is available is `glslangValidator`
which similarly can be invoked to compile GLSL shaders -> SPIR-V. It's nice to use
this because Homebrew has it:

```
brew install glslang
sudo ln -s $(which glslangValidator) /usr/local/bin/glslc
```

However, `glslc --target-env=vulkan1.2` is not supported with an `=` between the flag
and value. `glslc --target-env vulkan1.2` works, though, so I am proposing this minor
change.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>